verify_delegate() with succinct hash delegation
Change Metadata.verify_delegate(), so it can handle delegated roles
where succinct hash delegation is used.
In the case of succinct hash delegation, when the delegator is "targets"
we have to make sure that a given role "r" contains a succinct hash
delegation information and that the "delegated_role" is a role name is a
bin handled by "r".
This complication comes from the idea that we could have two delegated
roles with succinct hash delegation information and there could be
other non-succinct roles as well.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>