Clean up ReadTimeoutError usage.
ReadTimeoutError is actually a urllib3 exception.
We need it because requests Response.raw exposes the urllib3 Response, from which we call the read() method.
The exception is imported from requests.adapters, instead of directly from urllib3.exceptions, to prevent an explicit dependency on urllib3.
This is deliberate, as we only need the urllib3 exception because requests uses urllib3 (if the requests dependency would ever change, so should we).

Signed-off-by: dennisvang <29799340+dennisvang@users.noreply.github.com>