POC: use securesystemslib SerializationMixin
securesystemslib provides a mixin with generic methods to read and
write to and from bytes and files, which are reusable by TUF's
Metadata class, but also by the generic DSSE Envelope in
securesystemslib, and the in-toto DSSE Envelope and in-toto
Metablock.

This patch demonstrates how the mixin can be used instead of TUF's
own `Metadata.[to|from]_[bytes|file]` methods.

**CAVEAT:**

The default de/serializers can no longer be hardcoded into the
method (as they were in `Metadata.[to|from]_[bytes|file]`). To
still support defaults, users of the mixin must implement
`_default_serializer` and `_default_deserializer` helper methods.
This feels a bit intransparent and over-engineered.

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>