release: Use PyPI Trusted Publishing
Instead of using the secret stored in environment secrets, allow the
publish action to use the OIDC identity to authenticate to pypi.org.
This repository/workflow/environment has been marked as a "Trusted
Publisher" in pypi.org: this means PyPI should give the publish action a
short lived token to use for publishing.

This enables #2370: but the secret should still be removed before
closing the issue (maybe after one successful release with Trusted
Publishing).

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>