ngclient: Add an optional offline mode
This commit adds an offline configuration toggle to ngclient that makes
the client operate without naking network connections:
* feature is useful for repositories with slow moving content where
  a client may run hundreds or thousands of times without the repository
  changing (example: sigstore root-signing)
* this is strictly speaking not spec compliant

This is still a work-in-progress:
* Tests are currently unimplemented or broken
* Current iteration allows local metadata to be expired
  (it's still being discussed whether this is reasonable)

This is all work done by folks mentioned below (I just rebased and
squashed). I did not preserve the commits as the history contains merge
mistakes and reverts that would not improve our git history. The
included test was not correct so was removed for now.

Co-Authored-By: Emile Baez <ebaezmunne@gmail.com>
Co-Authored-By: Emerson Rounds <65183924+emboman13@users.noreply.github.com>
Co-Authored-By: omartounsi7 <62721212+omartounsi7@users.noreply.github.com>
Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>